move project api hooks to /go/components/project
