+++ go/cmd/apiserver/BUILD.bazel
@@ -18,8 +18,8 @@ go_library(
         "//go/base/config:go_default_library",
         "//go/base/env:go_default_library",
         "//go/base/zapfx:go_default_library",
+        "//go/components/project/apihook:go_default_library",
         "//go/logging:go_default_library",
-        "//go/project/apihook:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_uber_go_tally//:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
