+++ go/components/project/apihook/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["apihook.go"],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/project/apihook",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/components/project/apihook",
     visibility = ["//visibility:public"],
     deps = [
         "//go/api:go_default_library",
